requirements.txt
altgraph==0.17.2
asgiref==3.7.2
distlib==0.3.7
Django==4.2.6
django-cors-headers==4.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
filelock==3.12.4
future==0.18.2
macholib==1.15.2
pip==23.3.1
platformdirs==3.11.0
PyJWT==2.8.0
pytz==2023.3.post1
setuptools==58.0.4
six==1.15.0
sqlparse==0.4.4
typing_extensions==4.7.1
virtualenv==20.24.6
wheel==0.37.0

# npm install
# npm install @react-navigation/native-stack     
# npx expo install react-native-screens react-native-safe-area-context
# npm install @react-navigation/native
# npx expo install -fixx
# npm install axios
# npm install 
# npm i nativewind
# npm i --dev tailwindcss